Submitted Project
So this version might need some explanation. I had been developing this game in Eclipse, but for the final stretch, I had switched to developing it directly from the console to make sure I could make the jar file independent of eclipse's premade file export. This commit, though it comes much later than the others, was written within a week of the previous commit